feat: create component RecipeBlock for home page
