+++ src/pages/homePage/HomePage.tsx
@@ -1,49 +1,47 @@
-import { Divider, Layout, Spin, Typography } from 'antd';
+import { Divider, Layout } from 'antd';
 import { useEffect } from 'react';
 import style from './style.module.css';
 import { Header } from '../../commonComponents/header';
-import { RecipeCardList } from '../../components/recipeCardList';
 import { DayRecipeCard } from '../../components/dayRecipeCard';
 import { useAppDispatch, useAppSelector } from '../../hooks';
 import { getRandomDessertRecipes } from '../../store/randomDessertRecipe.slice';
+import { RecipeBlock } from './parts/recipeBlock';
 
 const { Content, Footer } = Layout;
-const { Title } = Typography;
 
 export function HomePage() {
-  const { recipes, loading, error } = useAppSelector(
-    (state) => state.randomDessertRecipes
-  );
+  const {
+    recipes: dessertRecipes,
+    loading: dessertLoading,
+    error: dessertError,
+  } = useAppSelector((state) => state.randomDessertRecipes);
   const dispatch = useAppDispatch();
 
   useEffect(() => {
     dispatch(getRandomDessertRecipes({}));
   }, [dispatch]);
-  console.log('loading = ', loading);
-  console.log('error = ', error);
-  console.log('recipes = ', recipes);
+
   return (
     <Layout className={style.layout}>
       <Header />
       <Divider />
       <Content className={style.content}>
         <DayRecipeCard />
-        <Title level={3}>Sweet tooth</Title>
-        {loading && (
-          <div className={style.spinContainer}>
-            <Spin size="large" data-testid="spin" />
-          </div>
-        )}
-        {error && <p className={style.error}>{error}</p>}
-        {!!recipes.length && <RecipeCardList recipes={recipes} />}
+        <RecipeBlock
+          loading={dessertLoading}
+          error={dessertError}
+          recipes={dessertRecipes}
+          title="Sweet tooth"
+        />
+        {/*
         <Title level={3}>Vegeterian</Title>
-        {loading && (
+        {/* {loading && (
           <div className={style.spinContainer}>
             <Spin size="large" data-testid="spin" />
           </div>
         )}
         {error && <p className={style.error}>{error}</p>}
-        {!!recipes.length && <RecipeCardList recipes={recipes} />}
+        {!!recipes.length && <RecipeCardList recipes={recipes} />} */}
       </Content>
       <Footer />
     </Layout>
