+++ src/pages/homePage/parts/recipeBlock/index.ts
@@ -1,0 +1,1 @@
+export { RecipeBlock } from './RecipeBlock';
