+++ src/pages/homePage/style.module.css
@@ -5,15 +5,3 @@
 .content {
   padding: 0 var(--left-right-content-padding);
 }
-
-.spinContainer {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
-.error {
-  margin-top: 1rem;
-  color: var(--color-error);
-  text-align: center;
-}
